pyproject.toml
[project]
name = "mcp-grna-alignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "gradio[mcp]>=5.33.1",
]
